bug: bugfixes for serving assets
