+++ src/core/decorators/redirect.js
@@ -1,0 +1,10 @@
+"use strict";
+
+module.exports = function(target) {
+  if (typeof target === "string") {
+    this.writeHead(303, {
+      Location: target
+    });
+    this.end();
+  }
+};
